Add FirebaseUI Authentication
Add a sign-in button and a user icon to the top left. When the user
is signed in, the button is not displayed. When the user is
signed out, the icon is not displayed. Clicking on the sign-in
button pops up a gmail oauth modal which allows user to sign in.